requirements.txt
Django==1.9.7
django-ckeditor==5.0.3
django-crispy-forms==1.6.0
django-filebrowser==3.7.1
django-filer==1.2.2
django-grappelli==2.8.1
django-mobile==0.7.0
django-mptt==0.8.4
django-multiupload==0.5.1
django-mutant==0.2.1
django-picklefield==0.3.2
django-polymodels==1.4.5
django-polymorphic==0.8.1
django-registration-redux==1.4
easy-thumbnails==2.3
patool==1.12
Pillow==3.3.0
Unidecode==0.4.19
Whoosh==2.7.4
